Module tmc_langs_framework::domain

Structs

ExerciseDesc

ExercisePackagingConfiguration

RunResult

TestDesc

Functions

prepare_solutions
Walks through each given path, processing files and copying them into the destination.

prepare_stubs
Walks through each given path, processing files and copying them into the destination.