Enum tmc_langs_framework::ExitStatus

pub enum ExitStatus {
    Exited(u32),
    Signaled(u8),
    Other(i32),
    Undetermined,
}

Exit status of a process.

Variants

Exited(u32)

The process exited with the specified exit code.

Note that the exit code is limited to a much smaller range on most platforms.

Signaled(u8)

The process exited due to a signal with the specified number.

This variant is never created on Windows, where signals of Unix kind do not exist.

Other(i32)

The process exit status cannot be described by the preceding two variants.

This should not occur in normal operation.

Undetermined

It is known that the process has completed, but its exit status is unavailable.

This should not occur in normal operation, but is possible if for example some foreign code calls waitpid() on the PID of the child process.

Implementations

impl ExitStatus

pub fn success(self) -> bool

True if the exit status of the process is 0.

pub fn is_killed_by<T>(self, signum: T) -> bool
where T: Eq + From<u8>,

True if the subprocess was killed by a signal with the specified number.

You can pass the concrete libc signal numbers to this function, such as status.is_killed_by(libc::SIGABRT).