Module tmc_langs_framework::io::submission_processing

Functions for processing submissions.

Functions

contains_tmcignore

is_hidden_dir

move_files
Moves some of the contents of source to target based on the given policy. For example, a file source/foo.java would be moved to target/foo.java.

prepare_solutions
Walks through each given path, processing files and copying them into the destination.

prepare_stub
Walks through each given path, processing files and copying them into the destination.