Trait tmc_langs_framework::policy::StudentFilePolicy

pub trait StudentFilePolicy {
    fn get_config_file_parent_path(&self) -> &Path;

    fn is_student_source_file(&self, file_path: &Path) -> bool;

    fn is_student_file(
        &self, 
        file_path: &Path, 
        project_root_path: &Path, 
        tmc_project_yml: &TmcProjectYml
    ) -> Result<bool, TmcError> { ... }

    fn get_tmc_project_yml(&self) -> Result<TmcProjectYml, TmcError> { ... }

    fn is_extra_student_file(
        &self, 
        file_path: &Path, 
        tmc_project_yml: &TmcProjectYml
    ) -> Result<bool, TmcError> { ... }

    fn is_updating_forced(
        &self, 
        path: &Path, 
        tmc_project_yml: &TmcProjectYml
    ) -> Result<bool, TmcError> { ... }
}

Specifies which files are student files.

Student files are any files that are expected to be modified and/or created by the student. That is, any files that should not be overwritten when when updating an already downloaded exercise and any files that should be submitted to the server.

Required methods

fn get_config_file_parent_path(&self) -> &Path

fn is_student_source_file(&self, file_path: &Path) -> bool

Checks whether the file is a student source file. The file_path can be assumed to be a relative path starting from the project root directory.

Loading content...
Provided methods

fn is_student_file(
&self,
file_path: &Path,
project_root_path: &Path,
tmc_project_yml: &TmcProjectYml
) -> Result<bool, TmcError>

Determines whether a file is a student source file.

A file should be considered a student source file if it resides in a location the student is expected to create his or her own source files in the general case. Any special cases are specified as ExtraStudentFiles in a separate configuration.

For example in a Java project that uses Apache Ant, should return true for any files in the src directory.

The file_path should be some file inside project_root_path.

Errors

Returns an error if either the file_path or project_root_path don't exist, or if file_path is not in project_root_path.

fn get_tmc_project_yml(&self) -> Result<TmcProjectYml, TmcError>

fn is_extra_student_file(
&self,
file_path: &Path,
tmc_project_yml: &TmcProjectYml
) -> Result<bool, TmcError>

Determines whether a file is an extra student file.

The file_path should be relative, starting from the project root.

fn is_updating_forced(
&self,
path: &Path,
tmc_project_yml: &TmcProjectYml
) -> Result<bool, TmcError>

Used to check for files which should always be overwritten.

The file_path should be relative, starting from the project root.

Loading content...