Struct tmc_langs_framework::RunResult

pub struct RunResult {
    pub status: RunStatus,
    pub test_results: Vec<TestResult>,
    pub logs: HashMap<String, String>,
}

The result of running an exercise’s test suite against a submission.

Fields
status: RunStatus

The overall status of a test run.

test_results: Vec<TestResult>

Whether each test passed and which points were awarded.

logs: HashMap<String, String>

Logs from the test run. The key may be an arbitrary string identifying the type of log.

Implementations

impl RunResult

pub fn new(
status: RunStatus,
test_results: Vec<TestResult>,
logs: HashMap<String, String>
) -> Self