Crate tmc_langs_framework

Contains functionality for dealing with projects.

Re-exports

pub use self::error::TmcError;

pub use self::plugin::LanguagePlugin;

pub use self::policy::StudentFilePolicy;

pub use self::tmc_project_yml::TmcProjectYml;

pub use anyhow;

pub use nom;

pub use subprocess;

pub use zip;

Modules

command

Custom wrapper for Command that supports timeouts and contains custom error handling.

domain

Contains structs that model data related to exercises.

error

file_util

Various utility functions, primarily wrapping the standard library’s IO and filesystem functions

meta_syntax

Contains utilities for parsing annotated exercise source files, separating lines into strings, stubs and solutions so that they can be more easily filtered later.

plugin

Contains LanguagePlugin.

policy

Contains StudentFilePolicy.

tmc_project_yml

Contains a struct that models the .tmcproject.yml file.

Macros

lock