Module tmc_langs_framework::domain

Contains structs that model data related to exercises.

Modules

meta_syntax
Contains utilities for parsing annotated exercise source files, separating lines into strings, stubs and solutions so that they can be more easily filtered later.

Structs

ExerciseDesc
A description of an exercise.

ExercisePackagingConfiguration
Represents configuration based on which submission may be packaged.

NoTests

NoTestsPoints

PythonVer

RunResult
The result of running an exercise's test suite against a submission.

TestDesc
A description of an exercise's test case.

TestResult
The result of a single test case.

TmcProjectYml
Extra data from a .tmcproject.yml file.

ValidationError

ValidationResult

Enums

IntOrString

NoTestsWrapper

RunStatus
The overall status of a test run.

Strategy