Trait tmc_langs_framework::plugin::LanguagePlugin

pub trait LanguagePlugin {
    type StudentFilePolicy: StudentFilePolicy + 'static;

    const PLUGIN_NAME: &'static str;

    fn scan_exercise(
        &self, 
        path: &Path, 
        exercise_name: String
    ) -> Result<ExerciseDesc>;

    fn run_tests_with_timeout(
        &self, 
        path: &Path, 
        timeout: Option<Duration>
    ) -> Result<RunResult>;

    fn get_student_file_policy(project_path: &Path) -> Self::StudentFilePolicy;

    fn find_project_dir_in_zip<R: Read + Seek>(
        zip_archive: &mut ZipArchive<R>
    ) -> Result<PathBuf>;

    fn is_exercise_type_correct(path: &Path) -> bool;

    fn clean(&self, path: &Path) -> Result<()>;

    fn find_exercises(&self, base_path: &Path) -> Vec<PathBuf> { ... }

    fn run_tests(&self, path: &Path) -> Result<RunResult> { ... }

    fn prepare_submission(
        &self, 
        policy: Self::StudentFilePolicy, 
        submission_path: &Path, 
        dest_path: &Path
    ) -> Result<()> { ... }

    fn prepare_stub(
        &self, 
        exercise_path: &Path, 
        repo_path: &Path, 
        dest_path: &Path
    ) -> Result<()> { ... }

    fn prepare_solution(
        &self, 
        exercise_paths: Vec<PathBuf>, 
        dest_path: &Path
    ) -> Result<()> { ... }

    fn check_code_style(
        &self, 
        _path: &Path, 
        _locale: Language
    ) -> Option<ValidationResult> { ... }

    fn compress_project(&self, path: &Path) -> Result<Vec<u8>> { ... }

    fn extract_project(
        &self, 
        compressed_project: &Path, 
        target_location: &Path, 
        clean: bool
    ) -> Result<()> { ... }

    fn maybe_copy_shared_stuff(&self, dest_path: &Path) -> Result<()> { ... }

    fn get_exercise_packaging_configuration(
        &self, 
        path: &Path
    ) -> Result<ExercisePackagingConfiguration> { ... }

    fn get_default_student_file_paths(&self) -> Vec<PathBuf> { ... }

    fn get_default_exercise_file_paths(&self) -> Vec<PathBuf> { ... }
}

The trait that each language plug-in must implement.

These implement the operations needed by the TMC server to support a programming language. These are provided as a library to IDE plug-ins as a convenience. IDE plug-ins often need additional integration work to support a language properly. This interface does NOT attempt to provide everything that an IDE plug-in might need to fully support a language.

Parts of this interface may be called in a TMC sandbox.

Implementations must be thread-safe and preferably fully stateless. Users of this interface are free to cache results if needed.

Associated Types

type StudentFilePolicy: StudentFilePolicy + 'static

Loading content...
Associated Constants

const PLUGIN_NAME: &'static str

Loading content...
Required methods

fn scan_exercise(
&self,
path: &Path,
exercise_name: String
) -> Result<ExerciseDesc>

Produces an exercise description of an exercise directory.

This involves finding the test cases and the points offered by the exercise.

Must return Err if the given path is not a valid exercise directory for this language.

fn run_tests_with_timeout(
&self,
path: &Path,
timeout: Option<Duration>
) -> Result<RunResult>

Runs the tests for the exercise with the given timeout.

fn get_student_file_policy(project_path: &Path) -> Self::StudentFilePolicy

fn find_project_dir_in_zip<R: Read + Seek>(
zip_archive: &mut ZipArchive<R>
) -> Result<PathBuf>

Searches the zip for a valid project directory. Note that the returned path may not actually have an entry in the zip.

fn is_exercise_type_correct(path: &Path) -> bool

Tells if there's a valid exercise in this path.

fn clean(&self, path: &Path) -> Result<()>

Runs clean command e.g make clean for make or mvn clean for maven.

Loading content...
Provided methods

fn find_exercises(&self, base_path: &Path) -> Vec<PathBuf>

Returns a list of all directories inside that contain an exercise in this language.

These directories might overlap with directories returned by some other language plug-in.

fn run_tests(&self, path: &Path) -> Result<RunResult>

Runs the tests for the exercise.

fn prepare_submission(
&self,
policy: Self::StudentFilePolicy,
submission_path: &Path,
dest_path: &Path
) -> Result<()>

Prepares a submission for processing in the sandbox.

The destination path is initialised with the original exercise as it appears in the course repository. The implementation should copy over a selection of files from the submission so that the student cannot e.g. easily replace the tests.

fn prepare_stub(
&self,
exercise_path: &Path,
repo_path: &Path,
dest_path: &Path
) -> Result<()>

Prepares a stub exercise from the original.

The stub is a copy of the original where the model solution and special comments have been stripped and stubs like ('return 0') have been added.

fn prepare_solution(
&self,
exercise_paths: Vec<PathBuf>,
dest_path: &Path
) -> Result<()>

Prepares a presentable solution from the original.

The solution usually has stubs and special comments stripped.

fn check_code_style(
&self,
_path: &Path,
_locale: Language
) -> Option<ValidationResult>

Run checkstyle or similar plugin to project if applicable, empty by default

fn compress_project(&self, path: &Path) -> Result<Vec<u8>>

Compress a given project so that it can be sent to the TestMyCode server.

fn extract_project(
&self,
compressed_project: &Path,
target_location: &Path,
clean: bool
) -> Result<()>

Extract a given archive file containing a compressed project to a target location.

This will overwrite any existing files as long as they are not specified as student files by the language dependent student file policy.

fn maybe_copy_shared_stuff(&self, dest_path: &Path) -> Result<()>

Copy shared stuff to stub or solution used for example for copying tmc-junit-runner.

fn get_exercise_packaging_configuration(
&self,
path: &Path
) -> Result<ExercisePackagingConfiguration>

Returns configuration which is used to package submission on tmc-server.

fn get_default_student_file_paths(&self) -> Vec<PathBuf>

fn get_default_exercise_file_paths(&self) -> Vec<PathBuf>

Loading content...