Trait tmc_langs_framework::policy::StudentFilePolicy

pub trait StudentFilePolicy {
    fn get_config_file_parent_path(&self) -> &Path;

    fn is_student_source_file(&self, path: &Path) -> bool;

    fn is_student_file(
        &self, 
        path: &Path, 
        project_root_path: &Path
    ) -> Result<bool> { ... }

    fn is_extra_student_file(&self, path: &Path) -> Result<bool> { ... }

    fn is_updating_forced(&self, path: &Path) -> Result<bool> { ... }
}

Specifies which files are student files.

Student files are any files that are expected to be modified and/or created by the student. That is, any files that should not be overwritten when when updating an already downloaded exercise and any files that should be submitted to the server.

Required methods

fn get_config_file_parent_path(&self) -> &Path

fn is_student_source_file(&self, path: &Path) -> bool

Loading content...
Provided methods

fn is_student_file(&self, path: &Path, project_root_path: &Path) -> Result<bool>

Determines whether a file is a student source file.

A file should be considered a student source file if it resides in a location the student is expected to create his or her own source files in the general case. Any special cases are specified as ExtraStudentFiles in a separate configuration.

For example in a Java project that uses Apache Ant, should return true for any files in the src directory.

fn is_extra_student_file(&self, path: &Path) -> Result<bool>

Determines whether a file is an extra student file.

fn is_updating_forced(&self, path: &Path) -> Result<bool>

Used to check for files which should always be overwritten.

Loading content...