tmc_langs_framework

Struct TmcCommand

pub struct TmcCommand { /* private fields */ }

Wrapper around subprocess::Exec

Implementations

impl TmcCommand

pub fn new(cmd: impl AsRef<OsStr>) -> Self

Creates a new command

pub fn piped(cmd: impl AsRef<OsStr>) -> Self

Creates a new command with piped stdout/stderr.

pub fn with(self, f: impl FnOnce(Exec) -> Exec) -> Self

Allows modification of the internal command without providing access to it.

pub fn set_stdin_data(self, data: String) -> Self

Gives the command data to write into stdin.

pub fn status(self) -> Result<ExitStatus, TmcError>

Executes the command and waits for its output.

pub fn output(self) -> Result<Output, TmcError>

Executes the command and waits for its output.

pub fn output_checked(self) -> Result<Output, TmcError>

Executes the command and waits for its output and errors if the status is not successful.

pub fn output_with_timeout(self, timeout: Duration) -> Result<Output, TmcError>

Executes the command and waits for its output with the given timeout.

pub fn output_with_timeout_checked( self, timeout: Duration, ) -> Result<Output, TmcError>

Executes the command and waits for its output with the given timeout and errors if the status is not successful.